Restructure ethermint/rpc repo structure and change import statements
