+++ rpc/apis.go
@@ -8,16 +8,16 @@ import (
 
 	"github.com/ethereum/go-ethereum/rpc"
 
-	"github.com/tharsis/ethermint/ethereum/rpc/backend"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/debug"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/eth"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/eth/filters"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/miner"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/net"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/personal"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/txpool"
-	"github.com/tharsis/ethermint/ethereum/rpc/namespaces/web3"
-	"github.com/tharsis/ethermint/ethereum/rpc/types"
+	"github.com/tharsis/ethermint/rpc/ethereum/backend"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/debug"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/eth"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/eth/filters"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/miner"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/net"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/personal"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/txpool"
+	"github.com/tharsis/ethermint/rpc/ethereum/namespaces/web3"
+	"github.com/tharsis/ethermint/rpc/ethereum/types"
 
 	rpcclient "github.com/tendermint/tendermint/rpc/jsonrpc/client"
 )
