+++ rpc/ethereum/namespaces/eth/filters/filter_system.go
@@ -22,7 +22,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/tharsis/ethermint/ethereum/rpc/pubsub"
+	"github.com/tharsis/ethermint/rpc/ethereum/pubsub"
 	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 )
