+++ rpc/ethereum/namespaces/eth/filters/filters.go
@@ -5,7 +5,7 @@ import (
 	"encoding/binary"
 	"math/big"
 
-	"github.com/tharsis/ethermint/ethereum/rpc/types"
+	"github.com/tharsis/ethermint/rpc/ethereum/types"
 
 	"github.com/pkg/errors"
 	"github.com/tendermint/tendermint/libs/log"
