+++ rpc/ethereum/namespaces/eth/filters/utils.go
