+++ rpc/ethereum/namespaces/txpool/api.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
 
-	"github.com/tharsis/ethermint/ethereum/rpc/types"
+	"github.com/tharsis/ethermint/rpc/ethereum/types"
 )
 
 // PublicAPI offers and API for the transaction pool. It only operates on data that is non-confidential.
