+++ rpc/ethereum/types/utils.go
