+++ server/json_rpc.go
@@ -11,7 +11,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/cosmos/cosmos-sdk/server/types"
 	ethrpc "github.com/ethereum/go-ethereum/rpc"
-	"github.com/tharsis/ethermint/ethereum/rpc"
+	"github.com/tharsis/ethermint/rpc/ethereum"
 
 	"github.com/tharsis/ethermint/server/config"
 )
