+++ testutil/network/util.go
@@ -30,7 +30,7 @@ import (
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
-	"github.com/tharsis/ethermint/ethereum/rpc"
+	"github.com/tharsis/ethermint/rpc/ethereum"
 	ethsrv "github.com/tharsis/ethermint/server"
 	ethermint "github.com/tharsis/ethermint/types"
 )
